+++ src/lib/server/db/actions/helpers/account/accountCreateInsertionData.ts
@@ -1,6 +1,6 @@
 import type { CreateAccountSchemaType } from '$lib/schema/accountSchema';
-import { statusUpdate } from './statusUpdate';
-import { updatedTime } from './updatedTime';
+import { statusUpdate } from '../statusUpdate';
+import { updatedTime } from '../updatedTime';
 import { combinedAccountTitleSplitRequired } from '$lib/helpers/combinedAccountTitleSplit';
 
 export const accountCreateInsertionData = (data: CreateAccountSchemaType, id: string) => {
