+++ src/lib/server/db/actions/helpers/category/categoryCreateInsertionData.ts
@@ -1,7 +1,7 @@
 import type { CreateCategorySchemaType } from '$lib/schema/categorySchema';
-import { statusUpdate } from './statusUpdate';
+import { statusUpdate } from '../statusUpdate';
 import { combinedTitleSplitRequired } from '$lib/helpers/combinedTitleSplit';
-import { updatedTime } from './updatedTime';
+import { updatedTime } from '../updatedTime';
 
 export const categoryCreateInsertionData = (data: CreateCategorySchemaType, id: string) => {
 	return {
