+++ src/lib/server/db/actions/helpers/label/labelCreateInsertionData.ts
@@ -1,6 +1,6 @@
 import type { CreateLabelSchemaType } from '$lib/schema/labelSchema';
-import { statusUpdate } from './statusUpdate';
-import { updatedTime } from './updatedTime';
+import { statusUpdate } from '../statusUpdate';
+import { updatedTime } from '../updatedTime';
 
 export const labelCreateInsertionData = (data: CreateLabelSchemaType, id: string) => {
 	return {
