+++ src/lib/server/db/actions/helpers/seed/seedTagData.ts
@@ -1,5 +1,5 @@
 import type { CreateTagSchemaType } from '$lib/schema/tagSchema';
-import { getRandomInteger } from './getRandom';
+import { getRandomInteger } from '../getRandom';
 
 export const createTag = (): CreateTagSchemaType => {
 	return {
