+++ src/lib/server/db/actions/helpers/tag/tagCreateInsertionData.ts
@@ -1,7 +1,7 @@
 import type { CreateTagSchemaType } from '$lib/schema/tagSchema';
-import { statusUpdate } from './statusUpdate';
+import { statusUpdate } from '../statusUpdate';
 import { combinedTitleSplitRequired } from '$lib/helpers/combinedTitleSplit';
-import { updatedTime } from './updatedTime';
+import { updatedTime } from '../updatedTime';
 
 export const tagCreateInsertionData = (data: CreateTagSchemaType, id: string) => {
 	return {
