+++ src/routes/(loggedIn)/labels/+page.server.ts
@@ -1,7 +1,7 @@
 import { authGuard } from '$lib/authGuard/authGuardConfig.js';
 import { serverPageInfo } from '$lib/routes.js';
 import { defaultJournalFilter } from '$lib/schema/journalSchema';
-import { labelFilterToText } from '$lib/server/db/actions/helpers/labelFilterToQuery.js';
+import { labelFilterToText } from '$lib/server/db/actions/helpers/label/labelFilterToQuery.js';
 import { tActions } from '$lib/server/db/actions/tActions';
 import { db } from '$lib/server/db/db';
 import { logging } from '$lib/server/logging';
